implemented vulkan into engine loop
