+++ crates/ris_core/Cargo.toml
@@ -6,8 +6,10 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-sdl2 = "0.34.5"
 chrono = "0.4"
+sdl2 = "0.34.5"
+vulkano = "0.33.0"
+shaderc = "0.8"
 ris_data = { path = "../ris_data" }
 ris_input = { path = "../ris_input" }
 ris_jobs = { path = "../ris_jobs" }
