+++ crates/ris_core/src/god_job.rs
@@ -5,7 +5,7 @@ use ris_data::gameloop::{
 use ris_jobs::job_system;
 
 use crate::{
-    gameloop::{logic_frame, output_frame},
+    gameloop::logic_frame,
     god_object::GodObject,
 };
 
@@ -36,12 +36,17 @@ pub fn run(mut god_object: GodObject) -> GameloopState {
 
         // submit jobs
         let output_future = job_system::submit(move || {
-            output_frame::run(
-                current_output,
+            let mut output_frame = god_object.output_frame;
+            let mut current_output = current_output;
+            let gameloop_state = output_frame
+                .run(
+                &mut current_output,
                 &previous_output_for_output,
                 &previous_logic_for_output,
                 &frame_for_output,
-            )
+            );
+
+            (output_frame, current_output, gameloop_state)
         });
 
         let logic_future = job_system::submit(move || {
@@ -60,9 +65,10 @@ pub fn run(mut god_object: GodObject) -> GameloopState {
 
         // wait for jobs
         let (new_logic_data, logic_state) = logic_future.wait();
-        let (new_output_data, output_state) = output_future.wait();
+        let (new_output_frame, new_output_data, output_state) = output_future.wait();
 
         // update buffers
+        god_object.output_frame = new_output_frame;
         current_input = new_input_data;
         current_logic = new_logic_data;
         current_output = new_output_data;
