+++ crates/ris_core/src/god_object.rs
@@ -2,10 +2,12 @@ use ris_data::info::app_info::AppInfo;
 use ris_video::video::Video;
 
 use crate::gameloop::input_frame::InputFrame;
+use crate::gameloop::output_frame::OutputFrame;
 
 pub struct GodObject {
     pub _video: Video,
     pub input_frame: InputFrame,
+    pub output_frame: OutputFrame,
     pub app_info: AppInfo,
 }
 
@@ -19,10 +21,12 @@ impl GodObject {
         let controller_subsystem = sdl_context.game_controller()?;
 
         let input_frame = InputFrame::new(event_pump, controller_subsystem);
+        let output_frame = OutputFrame::new()?;
 
         let god_object = GodObject {
             _video,
             input_frame,
+            output_frame,
             app_info,
         };
